% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.option.R
\name{add.option}
\alias{add.option}
\title{add.option}
\usage{
add.option(name, value, old.options, nesting.character = "\\\\.")
}
\arguments{
\item{name}{Option name. Nesting is indicated by character specified in nesting.character.}

\item{value}{New value of option}

\item{old.options}{Nested list the option should be added to}

\item{nesting.character}{String giving Regex pattern of nesting indication string. Defaults to '\\.'}
}
\value{
Nested list with updated options
}
\description{
Add option to nested list of options. Applied recursively
}
