% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{alternate.gene.sort}
\alias{alternate.gene.sort}
\title{alternate.gene.sort}
\usage{
alternate.gene.sort(coverage.statistics)
}
\arguments{
\item{coverage.statistics}{Data frame of coverage statistics}
}
\value{
Coverage statistics data frame sorted by alternating gene size
}
\description{
Given a data frame containing coverage statistics and gene information, returns that frame
 with the rows sorted by alternating gene size (for plotting)
}
\details{
Genes have varying numbers of associated amplicons and when plotting coverage statistics, 
 if two genes with very low numbers of amplicons are next to each other, the labels will overlap.
 This function sorts the coverage statistics data frame in a way that places the genes
 with the most amplicons (largest) next to those with the least (smallest).
}
