% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{build.variant.specification}
\alias{build.variant.specification}
\title{build.variant.specification}
\usage{
build.variant.specification(sample.ids, project.directory)
}
\arguments{
\item{sample.ids}{Vector of sample IDs. Must match subdirectories in project.directory.}

\item{project.directory}{Path to directory where sample subdirectories}
}
\value{
Data frame with paths to variant files.
}
\description{
Build data frame with paths to variant files.
}
\details{
Parses through sample IDs in a project directory and returns paths to variant files based on
	(theoretical) file name patterns. Useful for testing, or for entering the pipeline at non-traditional stages.
}
