% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.overlap.venn.diagram.R
\name{caller.overlap.venn.diagram}
\alias{caller.overlap.venn.diagram}
\title{. Make Venn diagram of variant caller overlap}
\usage{
caller.overlap.venn.diagram(variants, file.name)
}
\arguments{
\item{variants}{Data frame containing variants, typically from merge.variants function}

\item{file.name}{Name of output file}
}
\description{
. Make Venn diagram of variant caller overlap
}
