% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{capitalize.caller}
\alias{capitalize.caller}
\alias{capitalise.caller}
\title{capitalize.caller}
\usage{
capitalize.caller(caller)

capitalise.caller(caller)
}
\arguments{
\item{caller}{Character vector of callers to be capitalized}
}
\value{
Vector of same length as caller where eligible callers have been capitalized
}
\description{
Capitalize variant caller name
}
