% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{classify.variant}
\alias{classify.variant}
\title{classify.variant}
\usage{
classify.variant(ref, alt)
}
\arguments{
\item{ref}{Vector of reference bases}

\item{alt}{Vector of alternate bases}
}
\value{
Character vector giving type of variant.
}
\description{
Classify a variant as SNV, MNV, or indel based on the reference and alternative alleles
}
