% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.ides.output.R
\name{convert.ides.output}
\alias{convert.ides.output}
\title{Convert output of iDES step 1 to variant call format}
\usage{
convert.ides.output(filename, output = TRUE,
  output.suffix = ".calls.txt", minreads = 5, mindepth = 50)
}
\arguments{
\item{filename}{Path to file}

\item{output}{Logical indicating whether output should be saved to file. Defaults to true.}

\item{output.suffix}{Suffix to be appended to input filename if saving results to file}

\item{minreads}{Minimum numbers of reads}

\item{mindepth}{Minimum depth}
}
\value{
potential.calls Data frame of converted iDES calls
}
\description{
Convert output of iDES step 1 to variant call format
}
