% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create.directories}
\alias{create.directories}
\title{create.directories}
\usage{
create.directories(directory.names, path)
}
\arguments{
\item{directory.names}{Vector of names of directories to be created}

\item{path}{Path where directories should be created}
}
\description{
Create directories in a given path
}
