% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract.sample.ids}
\alias{extract.sample.ids}
\title{Extract sample IDs from file paths}
\usage{
extract.sample.ids(paths, from.filename = FALSE)
}
\arguments{
\item{paths}{vector of file paths}

\item{from.filename}{Logical indicating whether sample ID should be extracted from filename rather than path}
}
\value{
vector of extracted sample IDs
}
\description{
Extract sample IDs from a set of paths to files in sample-specific subfolders
}
