% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.variant.file.R
\name{filter.variant.file}
\alias{filter.variant.file}
\title{Filter variants in file.}
\usage{
filter.variant.file(variant.file, output.file, config.file = NULL,
  caller = c("vardict", "ides", "mutect", "pgm", "consensus"))
}
\arguments{
\item{variant.file}{Path to variant file}

\item{output.file}{Path to output file}

\item{config.file}{Path to config file to be used. If not supplied, will use the pre-existing VariTAS options.}

\item{caller}{Name of caller used (needed to match appropriate filters from settings)}
}
\value{
None
}
\description{
Filter variants from file, and save to output. Wrapper function that opens the 
variant file, calls filter.variants, and saves the result to file
}
