% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.variants.R
\name{filter.variants}
\alias{filter.variants}
\title{Filter variant calls}
\usage{
filter.variants(variants, caller = c("vardict", "ides", "mutect", "pgm",
  "consensus", "isis", "varscan", "lofreq"), config.file = NULL,
  verbose = FALSE)
}
\arguments{
\item{variants}{Data frame of variant calls with ANNOVAR annotation, or path to variant file.}

\item{caller}{Name of caller used (needed to match appropriate filters from settings)}

\item{config.file}{Path to config file to be used. If not supplied, will use the pre-existing VariTAS options.}

\item{verbose}{Logical indicating whether to output descriptions of filtering steps. Defaults to False, useful for debugging.}
}
\value{
filtered.variants Data frame of filtered variants
}
\description{
Filter data frame of variant calls based on thresholds specified in settings.
}
