% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fix.lofreq.af}
\alias{fix.lofreq.af}
\title{fix.lofreq.af}
\usage{
fix.lofreq.af(variant.specification)
}
\arguments{
\item{variant.specification}{Data frame of variant file information}
}
\description{
LoFreq also does not output allele frequencies, so this script calculates them from the
 DP (depth) and AD (variant allele depth) values--which are also not output nicely--
 and adds them to the annotated vcf.
}
