% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.names.R
\name{fix.names}
\alias{fix.names}
\title{Fix variant call column names}
\usage{
fix.names(column.names, variant.caller, sample.id = NULL)
}
\arguments{
\item{column.names}{Character vector of column names}

\item{variant.caller}{String giving name of variant caller}

\item{sample.id}{Optional sample ID. Used to fix headers.}
}
\value{
new.column.names Vector of column names after fixing]
}
\description{
Fix headers of variant calls to prepare for merging. This mostly 
consists in making sure the column headers will be unique by prefixing
the variant caller in question.
}
