% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fix.varscan.af}
\alias{fix.varscan.af}
\title{fix.varscan.af}
\usage{
fix.varscan.af(variant.specification)
}
\arguments{
\item{variant.specification}{Data frame of variant file information}
}
\description{
VarScan does not output allele frequencies, so this script calculates them from the
 DP (depth) and AD (variant allele depth) values and adds them to the annotated vcf.
}
