% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.base.substitution}
\alias{get.base.substitution}
\title{Get base substitution}
\usage{
get.base.substitution(ref, alt)
}
\arguments{
\item{ref}{Vector of reference bases}

\item{alt}{Vector of alternate bases}
}
\value{
base.substitutions
}
\description{
Get base substitution represented by pyrimidine in base pair. 
If more than one base in REF/ALT (i.e. MNV or indel rather than SNV), NA will be returned
}
