% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.bed.chromosomes}
\alias{get.bed.chromosomes}
\title{get.bed.chromosomes}
\usage{
get.bed.chromosomes(bed)
}
\arguments{
\item{bed}{Path to BED file}
}
\value{
Vector containing all chromosomes in BED file
}
\description{
Extract chromosomes from bed file
}
