% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.buildver}
\alias{get.buildver}
\title{get.buildver}
\usage{
get.buildver()
}
\value{
String giving reference genome build version (hg19 or hg38)
}
\description{
Get build version (hg19/hg38) based on settings.

Parses VariTAS pipeline settings to get the build version. When this function was first developed, the idea was to
be able to explicitly set ANNOVAR filenames based on the build version.
}
