% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.colours}
\alias{get.colours}
\title{Generate a colour scheme}
\usage{
get.colours(n)
}
\arguments{
\item{n}{Number of colours desired}
}
\value{
Colour.scheme generated colours
}
\description{
Generate a colour scheme
}
