% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.coverage.by.amplicon.R
\name{get.coverage.by.amplicon}
\alias{get.coverage.by.amplicon}
\title{Process sample coverage per amplicon data}
\usage{
get.coverage.by.amplicon(project.directory)
}
\arguments{
\item{project.directory}{Path to project directory. Each sample should have its own subdirectory}
}
\value{
combined.data Data frame giving coverage per amplicon per sample.
}
\description{
Parse coverageBed output to get coverage by amplicon
}
\references{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/coverage.html}
}
