% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.coverage.by.sample.statistics.R
\name{get.coverage.by.sample.statistics}
\alias{get.coverage.by.sample.statistics}
\title{Get statistics about coverage per sample}
\usage{
get.coverage.by.sample.statistics(project.directory)
}
\arguments{
\item{project.directory}{Path to project directory. Each sample should have its own subdirectory}
}
\value{
coverage.by.sample.statistics Data frame with coverage statistics per sample
}
\description{
Get statistics about coverage per sample
}
