% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.fasta.chromosomes}
\alias{get.fasta.chromosomes}
\title{get.fasta.chromosomes}
\usage{
get.fasta.chromosomes(fasta)
}
\arguments{
\item{fasta}{Path to reference fasta}
}
\value{
Vector containing all chromosomes in fasta file.
}
\description{
Extract chromosomes from fasta headers.
}
