% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.file.path}
\alias{get.file.path}
\title{get.file.path}
\usage{
get.file.path(sample.ids, directory, extension = NULL,
  allow.multiple = FALSE, allow.none = FALSE)
}
\arguments{
\item{sample.ids}{Vector of sample IDs to match filename on}

\item{directory}{Path to directory containing files}

\item{extension}{String giving extension of file}

\item{allow.multiple}{Boolean indicating whether to allow multiple matching files. 
Defaults to false, which throws an error if the query matches more than one file.}

\item{allow.none}{Boolean indicating whether to allow no matching files. 
Defaults to false, which throws an error if the query does not match any files.}
}
\value{
Paths to matched files
}
\description{
Get absolute path to sample-specific file for one or more samples
}
