% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.filters.R
\name{get.filters}
\alias{get.filters}
\title{get.filters}
\usage{
get.filters(filters)
}
\arguments{
\item{filters}{List of filter values. These will be updated to use default as the baseline, 
with caller-specific filters taking precedence if supplied.}
}
\value{
A list with updated filters
}
\description{
Determine filters per caller, given default and caller-specific values.
}
