% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.gene.R
\name{get.gene}
\alias{get.gene}
\title{get.gene}
\usage{
get.gene(bed.data)
}
\arguments{
\item{bed.data}{Data frame containing data from bed file}
}
\value{
vector of gene names, one entry for each row of \code{bed.data}
}
\description{
Use guesswork to extract gene from data frame of targeted panel data. The panel
 designer output can change, so try to guess what the format is.
}
