% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.miniseq.sample.files.R
\name{get.miniseq.sample.files}
\alias{get.miniseq.sample.files}
\title{get.miniseq.sample.files}
\usage{
get.miniseq.sample.files(sample.ids, directory,
  file.suffix = "_S\\\\d{1,2}_.*")
}
\arguments{
\item{sample.ids}{Vector of sample ids. Should form first part of file name}

\item{directory}{Directory where files can be found}

\item{file.suffix}{Regex expression for end of file name. For example, `file.suffix = '_S\\d{1,2}_.*_R1_.*'` will match R1 files.1 files.}
}
\value{
Character vector of file paths
}
\description{
Get files for a sample in a directory, ensuring there's only a single match per sample ID.
}
