% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.panel.coverage.by.gene}
\alias{get.panel.coverage.by.gene}
\title{Summarise panel coverage by gene}
\usage{
get.panel.coverage.by.gene(panel.file, gene.col = 5)
}
\arguments{
\item{panel.file}{path to panel}

\item{gene.col}{index of column containing gene name}
}
\value{
panel.coverage.by.gene data frame giving the number of amplicons and their total length by gene
}
\description{
Summarise panel coverage by gene
}
