% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.pool.from.panel.data}
\alias{get.pool.from.panel.data}
\title{Get pool corresponding to each amplicon}
\usage{
get.pool.from.panel.data(panel.data)
}
\arguments{
\item{panel.data}{data frame pool should be extracted from}
}
\value{
pools vector of pool information
}
\description{
The bed files are not consistent, so it's not clear where the pool will appear. 
This function parses through the columns to identify where the pool
}
