% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.varitas.options.R
\name{get.varitas.options}
\alias{get.varitas.options}
\title{Return VariTAS settings}
\usage{
get.varitas.options(option.name = NULL, nesting.character = "\\\\.")
}
\arguments{
\item{option.name}{Optional name of option. If no name is supplied, the full list of VariTAS options will be provided.}

\item{nesting.character}{String giving Regex pattern of nesting indication string. Defaults to '\\.'}
}
\value{
varitas.options list specifying VariTAS options
}
\description{
Return VariTAS settings
}
\examples{
reference.build <- get.varitas.options('reference_build');
mutect.filters <- get.varitas.options('filters.mutect');

}
