% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.vcf.chromosomes}
\alias{get.vcf.chromosomes}
\title{get.vcf.chromosomes}
\usage{
get.vcf.chromosomes(vcf)
}
\arguments{
\item{vcf}{Path to VCF file}
}
\value{
Vector containing all chromosomes in VCF
}
\description{
Extract chromosomes from a VCF file.
}
