% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.varitas.options.R
\name{in.varitas.options}
\alias{in.varitas.options}
\title{Check if a key is in VariTAS options}
\usage{
in.varitas.options(option.name = NULL, varitas.options = NULL,
  nesting.character = "\\\\.")
}
\arguments{
\item{option.name}{String giving name of option (with different levels joined by \code{nesting.character})}

\item{varitas.options}{Ampliseq options as a list. If missing, they will be obtained from \code{get.varitas.options()}}

\item{nesting.character}{String giving Regex pattern of nesting indication string. Defaults to '\\.'}
}
\value{
in.options Boolean indicating if the option name exists in the current varitas options
}
\description{
Check if a key is in VariTAS options
}
