% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{logical.to.character}
\alias{logical.to.character}
\title{logical.to.character}
\usage{
logical.to.character(x)
}
\arguments{
\item{x}{Vector to be converted}
}
\value{
Character vector after converting TRUE/FALSE
}
\description{
Convert a logical vector to a T/F coded character vector. Useful for preventing unwanted T->TRUE nucleotide conversions
}
