% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make.command.line.call}
\alias{make.command.line.call}
\title{Make string with command line call from its individual components}
\usage{
make.command.line.call(main.command, options = NULL, flags = NULL,
  option.prefix = "--", option.separator = " ", flag.prefix = "--")
}
\arguments{
\item{main.command}{String or vector of strings giving main part of command (e.g. "python test.py" or c("python", "test.py"))}

\item{options}{Named vector or list giving options}

\item{flags}{Vector giving flags to include.}

\item{option.prefix}{String to preface all options. Defaults to "--"}

\item{option.separator}{String to separate options form their values. Defaults to a single space.}

\item{flag.prefix}{String to preface all flags. Defaults to "--"}
}
\value{
command string giving command line call
}
\description{
Make string with command line call from its individual components
}
