% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.variants.R
\name{merge.variants}
\alias{merge.variants}
\title{Merge variants}
\usage{
\method{merge}{variants}(variant.specification, apply.filters = TRUE,
  remove.structural.variants = TRUE,
  separate.consensus.filters = FALSE, verbose = FALSE)
}
\arguments{
\item{variant.specification}{Data frame containing details of file paths, sample IDs, and caller.}

\item{apply.filters}{Logical indicating whether to apply filters. Defaults to TRUE.}

\item{remove.structural.variants}{Logical indicating whether structural variants (including CNVs) should be removed. Defaults to TRUE.}

\item{separate.consensus.filters}{Logical indicating whether to apply different thresholds to variants called by more than one caller 
(specified under consensus in config file). Defaults to FALSE.}

\item{verbose}{Logical indicating whether to print information to screen}
}
\value{
Data frame
}
\description{
Merge variants from multiple callers and return a data frame of merged calls. By default filtering
	is also applied, although this behaviour can be turned off by setting apply.filters to FALSE.
}
