% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{parse.job.dependencies}
\alias{parse.job.dependencies}
\title{Parse job dependencies}
\usage{
parse.job.dependencies(dependencies)
}
\arguments{
\item{dependencies}{Job dependency strings to be parsed.}
}
\value{
parsed.dependencies Vector of job dependencies after reformatting.
}
\description{
Parse job dependencies to make the functions more robust to alternate inputs (e.g. people writing alignment instead of bwa)
}
