% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.amplicon.coverage.per.sample.R
\name{plot.amplicon.coverage.per.sample}
\alias{plot.amplicon.coverage.per.sample}
\title{plot.amplicon.coverage.per.sample}
\usage{
\method{plot}{amplicon.coverage.per.sample}(coverage.statistics,
  output.directory)
}
\arguments{
\item{coverage.statistics}{Data frame containing coverage per amplicon per sample, typically from \code{get.coverage.by.amplicon}.}

\item{output.directory}{Directory where per sample plots should be saved}
}
\value{
None
}
\description{
Create one scatterplot per sample, showing coverage per amplicon, and an additional plot giving the median
}
