% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coverage.by.genome.order.R
\name{plot.coverage.by.genome.order}
\alias{plot.coverage.by.genome.order}
\title{Plot amplicon coverage by genome order}
\usage{
\method{plot}{coverage.by.genome.order}(coverage.data)
}
\arguments{
\item{coverage.data}{data frame with results from bedtools coverage command}
}
\description{
Use values obtained by bedtools coverage to make a plot of coverage by genome order
}
