% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.coverage.by.sample.R
\name{plot.coverage.by.sample}
\alias{plot.coverage.by.sample}
\title{plot.coverage.by.sample}
\usage{
\method{plot}{coverage.by.sample}(coverage.sample, file.name,
  statistic = c("mean", "median"))
}
\arguments{
\item{coverage.sample}{Data frame of coverage data, typically from \code{get.coverage.by.sample.statistics}}

\item{file.name}{Name of output file}

\item{statistic}{Statistic to be plotted (mean or median)}
}
\value{
None
}
\description{
Make a barplot of coverage per sample
}
