% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ontarget.percent.R
\name{plot.ontarget.percent}
\alias{plot.ontarget.percent}
\title{plot.ontarget.percent}
\usage{
\method{plot}{ontarget.percent}(coverage.sample, file.name)
}
\arguments{
\item{coverage.sample}{Data frame of coverage data, typically from \code{get.coverage.by.sample.statistics}}

\item{file.name}{Name of output file}
}
\value{
None
}
\description{
Make a scatterplot of ontarget percent per sample
}
