% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.paired.percent.R
\name{plot.paired.percent}
\alias{plot.paired.percent}
\title{plot.paired.percent}
\usage{
\method{plot}{paired.percent}(coverage.sample, file.name)
}
\arguments{
\item{coverage.sample}{Data frame of coverage data, typically from \code{get.coverage.by.sample.statistics}}

\item{file.name}{Name of output file}
}
\value{
None
}
\description{
Make a barplot of percent paired reads per sample
}
