% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.vcf.specification.R
\name{prepare.vcf.specification}
\alias{prepare.vcf.specification}
\title{prepare.vcf.specification}
\usage{
prepare.vcf.specification(vcf.details, sample.id.column = 1,
  vcf.column = 2, job.dependency.column = NA, caller.column = NA)
}
\arguments{
\item{vcf.details}{Data frame containing details of VCF files}

\item{sample.id.column}{Identifier of column in \code{vcf.details} containing sample IDs (index or name)}

\item{vcf.column}{Identifier of column in \code{vcf.details} containing VCF file (index or name)}

\item{job.dependency.column}{Identifier of column in \code{vcf.details} containing job dependency (index or name)}

\item{caller.column}{Identifier of column in \code{vcf.details} containing caller (index or name)}
}
\value{
Properly formatted VCF details
}
\description{
Prepare VCF specification data frame for annotation
}
