% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.coverage.reports.R
\name{process.coverage.reports}
\alias{process.coverage.reports}
\title{Process coverageBed reports}
\usage{
process.coverage.reports(project.directory)
}
\arguments{
\item{project.directory}{Path to project directory. Each sample should have its own subdirectory}
}
\value{
final.statistics data frame of coverage statistics generated by parsing through coverage reports
}
\description{
Process the coverage reports generated by bedtools coverage tool.
}
