% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.sample.contamination.checks.R
\name{process.sample.contamination.checks}
\alias{process.sample.contamination.checks}
\title{Process sample contamination checks}
\usage{
process.sample.contamination.checks(project.directory)
}
\arguments{
\item{project.directory}{Path to project directory. Each sample should have its own subdirectory}
}
\value{
freemix.scores Data frame giving sample contamination (column freemix) score per sample.
}
\description{
Takes *selfSM reports generated by VerifyBamID during alignment, and returns a vector of freemix scores.
The freemix score is a sequence only estimate of sample contamination that ranges from 0 to 1.

Note: Targeted panels are often too small for this step to work properly.
}
\references{
\url{https://genome.sph.umich.edu/wiki/VerifyBamID}
}
