% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.total.coverage.statistics.R
\name{process.total.coverage.statistics}
\alias{process.total.coverage.statistics}
\title{Process total coverage statistics}
\usage{
process.total.coverage.statistics(project.directory)
}
\arguments{
\item{project.directory}{Path to project directory. Each sample should have its own subdirectory}
}
\value{
data frame with extracted statistics
}
\description{
Process reports generated by flagstat. 
Assumes reports for before and after off-target filtering have been written to the same file, with separating headers
}
