% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.all.calls.R
\name{read.all.calls}
\alias{read.all.calls}
\title{read.all.calls}
\usage{
read.all.calls(sample.ids, caller = c("vardict", "mutect", "pgm"),
  project.directory, patient.ids = NULL, apply.filters = TRUE,
  variant.file.pattern = NULL)
}
\arguments{
\item{sample.ids}{Vector giving sample IDs to process}

\item{caller}{String indicating which caller was used}

\item{project.directory}{Path to project directory}

\item{patient.ids}{Optional vector giving patient ID (or other group) corresponding to each sample}

\item{apply.filters}{Logical indicating whether filters specified in VariTAS options should be applied. Defaults to TRUE. !}

\item{variant.file.pattern}{Pattern indicating where the variant file can be found. Sample ID should be indicated by SAMPLE_ID}
}
\value{
combined.variant.calls Data frame with variant calls from all patients
}
\description{
Read all calls made with a certain caller
}
