% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ides.file.R
\name{read.ides.file}
\alias{read.ides.file}
\title{Read iDES output}
\usage{
read.ides.file(filename)
}
\arguments{
\item{filename}{path to file}
}
\value{
ides.data  data frame read from iDES output
}
\description{
Read output from iDES_step1.pl and return data frame
}
