% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.variant.calls.R
\name{read.variant.calls}
\alias{read.variant.calls}
\title{Read variant calls from file and format for ease of downstream analyses.}
\usage{
read.variant.calls(variant.file, variant.caller)
}
\arguments{
\item{variant.file}{Path to variant file.}

\item{variant.caller}{String indicating which variant caller was used. Needed to format the headers.}
}
\value{
variant.calls Data frame of variant calls
}
\description{
Read variant calls from file and format for ease of downstream analyses.
}
