% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read.yaml}
\alias{read.yaml}
\title{read.yaml}
\usage{
read.yaml(file.name)
}
\arguments{
\item{file.name}{Path to yaml file}
}
\value{
list containing contents of yaml file
}
\description{
Read a yaml file
}
\examples{
read.yaml(file.path(path.package('varitas'), 'config.yaml'))

}
