% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.all.scripts.R
\name{run.all.scripts}
\alias{run.all.scripts}
\title{Run all the generated bash scripts without HPC commands}
\usage{
run.all.scripts(output.directory, stages.to.run = c("alignment", "qc",
  "calling", "annotation", "merging"), variant.callers = NULL,
  quiet = FALSE)
}
\arguments{
\item{output.directory}{Main directory where all files should be saved}

\item{stages.to.run}{A character vector of all stages that need running}

\item{variant.callers}{A character vector of variant callers to run}

\item{quiet}{Logical indicating whether to print commands to screen rather than submit jobs. Defaults to FALSE,
can be useful to set to TRUE for testing.}
}
\value{
None
}
\description{
Run all the scripts generated by previous parts of the pipeline, without using HPC commands
}
