% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.filtering.txt.R
\name{run.filtering.txt}
\alias{run.filtering.txt}
\title{Run filtering on an ANNOVAR-annotated txt file}
\usage{
run.filtering.txt(variant.file, caller = c("consensus", "vardict",
  "ides", "mutect"), output.directory = NULL, output.filename = NULL,
  code.directory = NULL, log.directory = NULL, config.file = NULL,
  job.dependencies = NULL, job.group = NULL, quiet = FALSE)
}
\arguments{
\item{variant.file}{Path to variant file}

\item{caller}{String giving variant caller that was used (affects which filters were applied.}

\item{output.directory}{Path to output directory}

\item{output.filename}{Name of resulting VCF file (defaults to SAMPLE_ID.vcf)}

\item{code.directory}{Path to directory where code should be stored}

\item{log.directory}{Path to directory where log files should be stored}

\item{config.file}{Path to config file}

\item{job.dependencies}{Vector with names of job dependencies}

\item{job.group}{Group job should belong to}

\item{quiet}{Logical indicating whether to print command to screen rather than submit it to the system. Defaults to false, useful for debugging.}
}
\description{
Run filtering on an ANNOVAR-annotated txt file
}
