% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.ides.R
\name{run.ides}
\alias{run.ides}
\title{Run iDES}
\usage{
run.ides(project.directory, sample.id.pattern = "._S\\\\d+$",
  sample.ids = NULL, job.dependencies = NULL)
}
\arguments{
\item{project.directory}{Directory containing files}

\item{sample.id.pattern}{Regex pattern to match sample IDs}

\item{sample.ids}{Vector of sample IDs}

\item{job.dependencies}{Vector of job dependencies}
}
\value{
None
}
\description{
Run iDES
}
\details{
Run iDES step 1on each sample, to tally up calls by strand.
Files are output to a the sample subdirectory
}
\note{
Deprecated function for running iDES. 
 Follows previous development package without specification data frames
}
\references{
\url{https://cappseq.stanford.edu/ides/}
}
