% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.target.qc.sample.R
\name{run.target.qc.sample}
\alias{run.target.qc.sample}
\title{Get ontarget reads and run coverage quality control}
\usage{
run.target.qc.sample(bam.file, sample.id, output.directory = NULL,
  code.directory = NULL, log.directory = NULL, config.file = NULL,
  job.dependencies = NULL, job.name = NULL, job.group = NULL,
  quiet = FALSE)
}
\arguments{
\item{bam.file}{Path to BAM file}

\item{sample.id}{Sample ID for labelling}

\item{output.directory}{Path to output directory}

\item{code.directory}{Path to directory where code should be stored}

\item{log.directory}{Path to directory where log files should be stored}

\item{config.file}{Path to config file}

\item{job.dependencies}{Vector with names of job dependencies}

\item{job.name}{Name of job to be submitted}

\item{job.group}{Group job should belong to}

\item{quiet}{Logical indicating whether to print command to screen rather than submit it to the system. Defaults to false, useful for debugging.}
}
\description{
Get ontarget reads and run coverage quality control
}
