% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save.config}
\alias{save.config}
\title{save.config}
\usage{
save.config(output.file = NULL)
}
\arguments{
\item{output.file}{Path to output file. If NULL (default), the config file will be saved as a temporary file.}
}
\value{
Path to config file
}
\description{
Save current varitas config options to a temporary file, and return filename.
}
