% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.coverage.excel.R
\name{save.coverage.excel}
\alias{save.coverage.excel}
\title{Save coverage statistics to multi-worksheet Excel file.}
\usage{
save.coverage.excel(project.directory, file.name, overwrite = TRUE)
}
\arguments{
\item{project.directory}{Path to project directory}

\item{file.name}{Name of output file}

\item{overwrite}{Logical indicating whether to overwrite existing file if it exists.}
}
\value{
None
}
\description{
Save coverage statistics to multi-worksheet Excel file.
}
