% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.variants.excel.R
\name{save.variants.excel}
\alias{save.variants.excel}
\title{Save variants to Excel.}
\usage{
save.variants.excel(variants, file.name, filters = NULL,
  overwrite = TRUE)
}
\arguments{
\item{variants}{Data frame containing variants}

\item{file.name}{Name of output file}

\item{filters}{Optional list of filters to be saved}

\item{overwrite}{Logical indicating whether to overwrite exiting file if it exists. Defaults to TRUE for consistency with other R functions.}
}
\description{
Makes an Excel workbook with variant calls. If filters are provided, these will 
be saved to an additional worksheet within the same file.
}
