% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.varitas.options.R
\name{set.varitas.options}
\alias{set.varitas.options}
\title{Set options for varitas pipeline.}
\usage{
set.varitas.options(...)
}
\arguments{
\item{\dots}{options to set}
}
\value{
None
}
\description{
Set or overwrite options for the VariTAS pipeline. Nested options should be separated by a dot. 
	For example, to update the reference genome for grch38, use reference_genome.grch38
}
\examples{
\dontrun{
set.varitas.options(reference_build = 'grch38'); 	
	set.varitas.options(
	filters.mutect.min_normal_depth = 10,
	filters.vardict.min_normal_depth = 10
	);
}
}
