% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{split.on.column}
\alias{split.on.column}
\title{split.on.column}
\usage{
\method{split}{on.column}(dat, column, split.character)
}
\arguments{
\item{dat}{Data frame to be processed}

\item{column}{Name of column to split on}

\item{split.character}{Pattern giving character to split column on}
}
\value{
Data frame after splitting on column
}
\description{
Split data frame on a concatenated column.
}
