% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sum.dp4}
\alias{sum.dp4}
\title{sum.dp4}
\usage{
\method{sum}{dp4}(dp4.str)
}
\arguments{
\item{dp4.str}{String of DP4 values in the form "1234,1234,1234,1234"}
}
\description{
Simply calculates the depth of coverage of the variant allele given a string of
 DP4 values
}
