% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tabular.mean}
\alias{tabular.mean}
\title{tabular.mean}
\usage{
tabular.mean(values, frequencies, ...)
}
\arguments{
\item{values}{vector of values}

\item{frequencies}{frequency corresponding to each value}

\item{...}{Additional parameters passed to \code{sum}}
}
\value{
calculated mean
}
\description{
Calculate the mean of data in tabular format
}
