% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tabular.median}
\alias{tabular.median}
\title{tabular.median}
\usage{
tabular.median(values, frequencies, ...)
}
\arguments{
\item{values}{Vector of values}

\item{frequencies}{Frequency corresponding to each value}

\item{...}{Additional parameters passed to \code{sum}}
}
\value{
calculated median
}
\description{
Calculate the median of data in tabular format
}
