% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinucleotide.barplot.R
\name{trinucleotide.barplot}
\alias{trinucleotide.barplot}
\title{Make barplot of trinucleotide substitutions}
\usage{
trinucleotide.barplot(variants, file.name)
}
\arguments{
\item{variants}{Data frame with variants}

\item{file.name}{Name of output file}
}
\value{
None
}
\description{
Make barplot of trinucleotide substitutions
}
