% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.recurrence.barplot.R
\name{variant.recurrence.barplot}
\alias{variant.recurrence.barplot}
\title{Make barplot of variants per caller}
\usage{
variant.recurrence.barplot(variants, file.name)
}
\arguments{
\item{variants}{Data frame with variants}

\item{file.name}{Name of output file}
}
\value{
None
}
\description{
Make barplot of variants per caller
}
