% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.caller.barplot.R
\name{variants.caller.barplot}
\alias{variants.caller.barplot}
\title{Make barplot of variants per caller}
\usage{
variants.caller.barplot(variants, file.name, group.by = NULL)
}
\arguments{
\item{variants}{Data frame with variants}

\item{file.name}{Name of output file}

\item{group.by}{Optional grouping variable for barplot}
}
\value{
None
}
\description{
Make barplot of variants per caller
}
