% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.sample.barplot.R
\name{variants.sample.barplot}
\alias{variants.sample.barplot}
\title{Make barplot of variants per sample}
\usage{
variants.sample.barplot(variants, file.name)
}
\arguments{
\item{variants}{Data frame with variants}

\item{file.name}{Name of output file}
}
\value{
None
}
\description{
Make barplot of variants per sample
}
