% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.bam.specification.R
\name{verify.bam.specification}
\alias{verify.bam.specification}
\title{Check that sample specification data frame matches expected format, and that all files exist}
\usage{
verify.bam.specification(bam.specification)
}
\arguments{
\item{bam.specification}{Data frame containing columns sample.id and tumour.bam, and optionally a column normal.bam.}
}
\value{
None
}
\description{
Check that sample specification data frame matches expected format, and that all files exist
}
