% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.bwa.index.R
\name{verify.bwa.index}
\alias{verify.bwa.index}
\title{verify.bwa.index}
\usage{
verify.bwa.index(fasta.file, error = FALSE)
}
\arguments{
\item{fasta.file}{Fasta file to check}

\item{error}{Logical indicating whether to throw an (informative) error if verification fails}
}
\value{
index.files.exist Logical indicating if bwa index files were found (only returned if error set to FALSE)
}
\description{
Verify that bwa index files exist for a fasta file
}
