% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.fasta.index.R
\name{verify.fasta.index}
\alias{verify.fasta.index}
\title{verify.fasta.index}
\usage{
verify.fasta.index(fasta.file, error = FALSE)
}
\arguments{
\item{fasta.file}{Fasta file to check}

\item{error}{Logical indicating whether to throw an (informative) error if verification fails}
}
\value{
faidx.exists Logical indicating if fasta index files were found (only returned if error set to FALSE)
}
\description{
Verify that fasta index files exist for a given fasta file.
}
