% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.fastq.specification.R
\name{verify.fastq.specification}
\alias{verify.fastq.specification}
\title{Check that FASTQ specification data frame matches expected format, and that all files exist}
\usage{
verify.fastq.specification(fastq.specification, paired.end = FALSE,
  files.ready = FALSE)
}
\arguments{
\item{fastq.specification}{Data frame containing columns sample.id and reads, and optionally a column mates}

\item{paired.end}{Logical indicating whether paired end reads are used}

\item{files.ready}{Logical indicating if the files already exist on disk. If there are job dependencies, this should be set to FALSE.}
}
\value{
None
}
\description{
Check that FASTQ specification data frame matches expected format, and that all files exist
}
