% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.sequence.dictionary.R
\name{verify.sequence.dictionary}
\alias{verify.sequence.dictionary}
\title{verify.sequence.dictionary}
\usage{
verify.sequence.dictionary(fasta.file, error = FALSE)
}
\arguments{
\item{fasta.file}{Fasta file to check}

\item{error}{Logical indicating whether to throw an (informative) error if verification fails}
}
\value{
dict.exists Logical indicating if sequence dictionary files were found (only returned if error set to FALSE)
}
\description{
Verify that sequence dictionary exists for a fasta file.
}
