% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.varitas.options.R
\name{verify.varitas.options}
\alias{verify.varitas.options}
\title{Check against common errors in the VariTAS options.}
\usage{
verify.varitas.options(stages.to.run = c("alignment", "qc", "calling",
  "annotation", "merging"), variant.callers = c("mutect", "vardict",
  "ides", "varscan", "lofreq", "muse"), varitas.options = NULL)
}
\arguments{
\item{stages.to.run}{Vector indicating which stages should be run. Defaults to all possible stages. 
If only running a subset of stages, only checks corresponding to the desired stages are run}

\item{variant.callers}{Vector indicating which variant callers to run. Only used if calling is in \code{stages.to.run}.}

\item{varitas.options}{Optional file path or list of VariTAS options.}
}
\value{
None
}
\description{
Check against common errors in the VariTAS options before launching into pipeline
}
