% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.vcf.specification.R
\name{verify.vcf.specification}
\alias{verify.vcf.specification}
\title{verify.vcf.specification}
\usage{
verify.vcf.specification(vcf.specification)
}
\arguments{
\item{vcf.specification}{VCF specification data frame}
}
\value{
None
}
\description{
Verify that VCF specification data frame fits expected format
}
